requirements.txt
Package             Version
------------------- ---------
Django              4.0.6
django-crispy-forms 1.14.0
django-rest-swagger 2.2.0
djangorestframework 3.13.1
